+++ core/data/src/main/java/it/stamp/data/service/UserBootstrapService.kt
@@ -1,4 +1,4 @@
-package it.stamp.data.bootstrap
+package it.stamp.data.service
 
 import com.google.firebase.firestore.FirebaseFirestore
 import it.stamp.data.firestore.model.FirestoreGroup
